#Started_with_parent

+++ ParentsAbroad.Interfaces/Repositories/IBaseRepository.cs
@@ -8,6 +8,6 @@ namespace ParentsAbroad.Interfaces.Repositories
         Task<IList<T>> GetByAsync(Expression<Func<T, bool>> filter);
         Task<T> GetByIdAsync(long id);
         Task<T> SaveOrUpdateAsync(T entity);
-        Task<bool> DeleteAsync(int id);
+        Task<bool> DeleteAsync(long id);
     }
 }
